电商数据抓取/requirements.txt
langgraph
langchain-openai
langchain-anthropic
python-dotenv
fpdf2
pandas
numpy
tavily-python
requests
openai
langchain-core
langchain-community
streamlit
